# Conflicts:
#	app/src/main/java/com/qnet/qnetclient/applocal/ui/info_local/CambiarImagenLocalFragment.kt

+++ app/src/main/java/com/qnet/qnetclient/data/repo/FirebaseRepo.kt
@@ -85,7 +85,7 @@ class FirebaseRepo {
             "horario" to info.horario,
             "descripcion" to info.tipo,
             "informacion" to info.informacion,
-            "telefono" to info.telefono?.toInt(),
+            "telefono" to info.telefono?.toLong(),
             "queueNumber" to 0,
             "queuedPeople" to arrayListOf(null)
         )
@@ -202,6 +202,9 @@ class FirebaseRepo {
         val listData = mutableListOf<Model>()
 
         getLocalesReference().observeForever {
+            if(it.size==0){
+                mutableData.value = listData
+            }
             for (reference in it) {
                 db.document("locales/${reference.keyLocal}").get().addOnSuccessListener { result ->
 
@@ -312,6 +315,9 @@ class FirebaseRepo {
         var mutableReference = mutableListOf<References>()
         val listData = mutableListOf<Model>()
         getMisColasReference().observeForever{
+            if(it.size==0){
+                mutableData.value = listData
+            }
             for(reference in it)
             {
                 db.document("locales/${reference.keyLocal}").get().addOnSuccessListener {result ->
@@ -480,7 +486,9 @@ class FirebaseRepo {
 
         ref.putFile(uri).addOnSuccessListener {
             ref.downloadUrl.addOnSuccessListener {
-
+                deleteImage(it.toString()).observeForever{result->
+                    mutableData.value = result
+                }
             }
         }.addOnFailureListener{
             mutableData.value= false
@@ -492,20 +500,29 @@ class FirebaseRepo {
     private fun deleteImage(path: String?): LiveData<Boolean> {
         val mutableData = MutableLiveData<Boolean>()
         mAuth = FirebaseAuth.getInstance()
-        db.document("locales/${mAuth.currentUser?.uid}").get().addOnSuccessListener {
+        db.document("locales/${mAuth.currentUser?.uid}").get().addOnSuccessListener { it ->
             val image = it.getString("image")
-            var location = image?.split("%2F")
-            location = location?.get(1)?.split("?")
-            val ref= FirebaseStorage.getInstance().getReference("/images/${location?.get(0)}")
-            ref.delete().addOnSuccessListener {
+            if(image==null){
                 if (path != null) {
-                    changeData("image",path)
+                    changeData("image",path).observeForever{result->
+                        mutableData.value = result
+                    }
+                }
+            }else{
+                var location = image.split("%2F")
+                location = location[1].split("?")
+                val ref= FirebaseStorage.getInstance().getReference("/images/${location[0]}")
+                ref.delete().addOnSuccessListener {
+                    if (path != null) {
+                        changeData("image",path).observeForever{result->
+                            mutableData.value = result
+                        }
+                    }
                 }
             }
         }
 
         return mutableData
     }
-
 
 }
